requirements.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==3.0.2
defusedxml==0.7.1
Django==4.2.1
django-channels==0.7.0
django-cors-headers==4.1.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
msgpack-python==0.5.6
oauthlib==3.2.2
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.1.1
python-environ==0.4.54
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.1
urllib3==2.0.2
zope.interface==6.0
